% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_sim.R
\name{ce_sim}
\alias{ce_sim}
\title{Markov model cost-effectiveness simulation}
\usage{
ce_sim(pop, lambda, c_unit, e_unit, pdecr = NA, delta = 0.035)
}
\arguments{
\item{pop}{Number of individuals}

\item{lambda}{Transition probabilities}

\item{c_unit}{Unit costs per state}

\item{e_unit}{Health unit values per state}

\item{pdecr}{Linear utility decrease per state}

\item{delta}{Discount rate}
}
\description{
Markov model cost-effectiveness simulation
}
