% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_pop.R
\name{init_pop}
\alias{init_pop}
\title{Create initial population}
\usage{
init_pop(n_init, n_sim, tmax, n_interv = 2)
}
\arguments{
\item{n_sim}{Number of simulations (from jags)}

\item{tmax}{Time horizon}

\item{n_interv}{Number of interventions, including status-quo}

\item{pop}{Number of individuals [state, time, n_sim]}
}
\description{
Create initial population
}
